v2.0
You can now hide and show the LCD and GPU windows with the buttons at the top right!
You can now customize the address ranges of the VIA and GPU!
The GPU's resolution and array size is _whatever you want it to be_!
The fonts actually work now!
And **COLORS!!**